t03_02_exer4: completado

+++ T3/Tareas/t03_02_Fernandez_Lamas/exer4/src/App.java
@@ -39,7 +39,7 @@ public class App {
                 case "b":
                     System.out.println("Indica el índice del módulo:");
                     int indiceModulo = scanner.nextInt();
-                    int aprobados = 0;
+                    double aprobados = 0;
                     for (int i = 0; i < numEstudiantes; i++) {
                         if (notas[indiceModulo][i]>=5.0){
                             aprobados++;
